pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jobagent"
version = "1.0.0"
description = "AI-powered personal job application agent"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "JobAgent Contributors" }]
keywords = ["job-search", "ai", "automation", "linkedin", "anthropic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]
dependencies = [
    "anthropic>=0.40.0",
    "playwright>=1.44.0",
    "pyyaml>=6.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "twilio>=9.0.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
jobagent = "jobagent.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/jobagent"
Repository = "https://github.com/yourusername/jobagent"
Issues = "https://github.com/yourusername/jobagent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["jobagent*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]


[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=jobagent --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
omit = ["tests/*", "*/migrations/*"]
